Created a dynamic array class.
Wrote the functions; insert(), removeAt(), indexOf(), and print().

insert(): adds a new item to the array. If the array is full, it creates a new array twice the size and copies the former values. Then sets the reference between the former and the latter.

removeAt(): removes the element at the given index. Shifts the other values by 1 to the left.

indexOf(): returns the number of the index of a given element.

print(): prints the array elements using a for loop.